# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skeleton Laravel",
        "description": "All API Documentations",
        "contact": {
            "email": "prajwalmaharjan58@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        },
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/api/v1/guest/register": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "summary": "Register New User",
                "description": "Registers new users",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Register"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/guest/login": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "summary": "User Login",
                "description": "User login",
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Login"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/guest/register/verifyEmail": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Verifies Email",
                "description": "Verifies the email",
                "operationId": "Verifies Email",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "expires",
                                    "id",
                                    "hash",
                                    "signature"
                                ],
                                "properties": {
                                    "expires": {
                                        "type": "string",
                                        "example": "163877180"
                                    },
                                    "id": {
                                        "type": "string",
                                        "example": "16"
                                    },
                                    "hash": {
                                        "type": "string",
                                        "example": "a2675156297f4d2806e00f12d1aebe77b4492975"
                                    },
                                    "signature": {
                                        "type": "string",
                                        "example": "fc559b82958efe6f9817161b63a9b0541f39e33f46701ede7b4b9d20d1247beb"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation Successfull"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/guest/sendVerificationEmail": {
            "post": {
                "tags": [
                    "Guest"
                ],
                "summary": "Resend Verification Email",
                "description": "Resend verification email",
                "operationId": "Send Verification Email",
                "requestBody": {
                    "description": "required params",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "user@mail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operation Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/guest/pages/{slug}": {
            "get": {
                "tags": [
                    "Guest"
                ],
                "summary": "Get Page",
                "description": "Gets pages from slug",
                "operationId": "getPages",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Page Slug",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Login": {
                "title": "Login",
                "description": "User Login request body data",
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string",
                        "example": "email"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Login"
                }
            },
            "Register": {
                "title": "Register",
                "description": "User register request body",
                "properties": {
                    "first_name": {
                        "title": "First Name",
                        "description": "First Name",
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Last Name",
                        "type": "string",
                        "example": "Cena"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "example": "John@cena.com"
                    },
                    "phone": {
                        "title": "Phone Number",
                        "description": "Phone Number",
                        "type": "string",
                        "example": "John@cena.com"
                    },
                    "gender": {
                        "title": "Gender",
                        "description": "Gender",
                        "example": "male"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "example": "password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Register"
                }
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}